@@ -4,7 +4,7 @@ version: 0.0.8
 homepage: https://github.com/JacobWrenn/notustohtml
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   notus:
